Move production seeders to the production directory
