+++ src/seeders/production/20210112200315-production-add-careers.ts
@@ -1,5 +1,5 @@
 import { QueryInterface } from "sequelize";
-import { Environment } from "../config/Environment";
+import { Environment } from "../../config/Environment";
 
 const TABLE_NAME = "Careers";
