+++ src/seeders/production/20210112201204-production-add-secretary-settings.ts
@@ -1,6 +1,6 @@
-import { Secretary } from "../models/Admin/Interface";
+import { Secretary } from "$models/Admin/Interface";
 import { QueryInterface } from "sequelize";
-import { Environment } from "../config/Environment";
+import { Environment } from "../../config/Environment";
 
 const TABLE_NAME = "SecretarySettings";
